+++ product/src/main/java/store/_0982/product/batch/processor/StatusProcessor.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.stereotype.Component;
-import store._0982.product.batch.client.OrderClient;
+import store._0982.product.client.OrderClient;
 import store._0982.product.domain.GroupPurchase;
 import store._0982.product.domain.GroupPurchaseStatus;
