+++ product/src/main/java/store/_0982/product/batch/writer/StatusWriter.java
@@ -6,7 +6,7 @@ import org.springframework.batch.item.Chunk;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.stereotype.Component;
 import store._0982.common.dto.ResponseDto;
-import store._0982.product.batch.client.OrderClient;
+import store._0982.product.client.OrderClient;
 import store._0982.product.domain.GroupPurchase;
 import store._0982.product.domain.GroupPurchaseRepository;
 import store._0982.product.domain.GroupPurchaseStatus;
