+++ product/src/main/java/store/_0982/product/client/OrderClient.java
@@ -1,7 +1,6 @@
-package store._0982.product.batch.client;
+package store._0982.product.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,13 +14,13 @@ import java.util.UUID;
 )
 public interface OrderClient {
 
-    @PostMapping("/api/orders/internal/{purchaseId}/status")
+    @PostMapping("/internal/orders/{purchaseId}/status")
     ResponseDto<Void> updateOrderStatus(
             @PathVariable("purchaseId") UUID purchaseId,
             @RequestParam("status") String status
     );
 
-    @PostMapping("/api/orders/internal/{purchaseId}/return")
+    @PostMapping("/internal/orders/{purchaseId}/return")
     ResponseDto<Void> returnOrders(
             @PathVariable("purchaseId") UUID purchaseId
     );
